+++ packages/apollo-language-server/src/engine/operations/checkSchema.ts
@@ -1,27 +1,4 @@
-import { IntrospectionSchema } from "graphql";
 import gql from "graphql-tag";
-
-export interface GitContextInput {
-  commit: string;
-  commiter?: string;
-  message?: string;
-  branch?: string;
-  remoteUrl?: string;
-}
-export interface HistoricQueryParameters {
-  from?: number;
-  to?: number;
-  queryCountThreshold?: number;
-  queryCountThresholdPercentage?: number;
-}
-export interface CheckSchemaVariables {
-  id: string;
-  schema: IntrospectionSchema;
-  tag?: string;
-  gitContext?: GitContextInput;
-  historicParameters?: HistoricQueryParameters;
-  frontend?: string;
-}
 
 export const CHECK_SCHEMA = gql`
   mutation CheckSchema(