+++ packages/apollo-language-server/src/engine/operations/uploadSchema.ts
@@ -1,14 +1,4 @@
-import { IntrospectionSchema } from "graphql";
 import gql from "graphql-tag";
-
-import { GitContextInput } from "./checkSchema";
-
-export interface UploadSchemaVariables {
-  id: string;
-  schema: IntrospectionSchema;
-  tag: string;
-  gitContext: GitContextInput;
-}
 
 export const UPLOAD_SCHEMA = gql`
   mutation UploadSchema(